gtao: fix banded false occlusion on distant large grounds
Two scale pathologies painted wide iso-depth AO bands across big planes
(rooms/scenes hundreds of units across; discriminator: --no-ssao removed
the bands):

- The aux G-buffer stored linear view-Z in half floats; fp16's ulp at
  z=7000 is ~8 world units, so a receding ground staircases and the AO
  height test flips per step. The aux buffer (MSAA attachment and resolve
  target) is now full float.

- When the AO's world radius projects below one half-res depth texel,
  every NEAREST tap returns the center texel's own Z, and on a tilted
  surface the flat-Z sample vector fabricates a phantom step-up occluder;
  whether taps escape the texel cycles with perspective, banding the
  result. Sub-resolution pixels now early-out as unoccluded -- below the
  sampling resolution no occlusion is measurable.

Verified: room.fbx ground clean at scene radius 2843; raiden unchanged
(its screen radius is far above one texel, the guard never triggers).

+++ cetra/src/engine.c
@@ -346,12 +346,14 @@ static int _create_msaa_attachments(Engine* engine, int rw, int rh, int samples)
 
     // 0: HDR color. 1: view-space normals (.xyz) for SSAO/SSR. 2: aux G-buffer --
     // motion .xy (TAA) + linear view-Z .z (GTAO). 3: base color .rgb + metalness
-    // .a for SSGI. RGBA16F everywhere the resolve blit needs float; albedo is LDR.
+    // .a for SSGI. Albedo is LDR. The aux buffer is FULL float: half floats
+    // quantize linear Z into scene-scale steps on large scenes (fp16 ulp at
+    // z=7000 is ~8 units) and GTAO reads the staircase as banded occlusion.
     _add_msaa_color_attachment(&engine->multisample_texture, GL_RGBA16F, GL_COLOR_ATTACHMENT0, rw,
                                rh, samples);
     _add_msaa_color_attachment(&engine->normal_multisample_texture, GL_RGBA16F, GL_COLOR_ATTACHMENT1,
                                rw, rh, samples);
-    _add_msaa_color_attachment(&engine->aux_multisample_texture, GL_RGBA16F, GL_COLOR_ATTACHMENT2,
+    _add_msaa_color_attachment(&engine->aux_multisample_texture, GL_RGBA32F, GL_COLOR_ATTACHMENT2,
                                rw, rh, samples);
     _add_msaa_color_attachment(&engine->albedo_multisample_texture, GL_RGBA8, GL_COLOR_ATTACHMENT3,
                                rw, rh, samples);
